replace update and bug fix
* replace off implemented
* file created has the permissions changed to '777'
* a few Tests have been updated because the output from generators has changed

+++ src/main/java/org/openmainframeproject/cobolcheck/services/cobolLogic/replace/Replace.java
@@ -1,5 +1,7 @@
 package org.openmainframeproject.cobolcheck.services.cobolLogic.replace;
 
+import org.openmainframeproject.cobolcheck.services.Config;
+import org.openmainframeproject.cobolcheck.services.filehelpers.FilePermission;
 import org.openmainframeproject.cobolcheck.services.log.Log;
 import org.openmainframeproject.cobolcheck.services.log.LogLevel;
 
@@ -84,7 +86,6 @@ public class Replace {
         String replacesString = source;
 
         for (ReplaceSet replaceSet : replaceMap) {
-            Log.trace("Replace.replace(): Key: <" + replaceSet.getFrom() + ">, Value: <" + replaceSet.getTo() + ">");
             replacesString = replaceSet.replaceInline(replacesString, lineNumber);
             if ((Log.level() == LogLevel.TRACE) && (!replacesString.equals(source))) {
                 Log.trace("Replace.replace(): Key: <" + replaceSet.getFrom() + ">, result: " + replacesString);
@@ -99,7 +100,6 @@ public class Replace {
 
 
     public static void inspectProgram(File cobolProgram) {
-        Log.trace("Replace.inspectProgram(): Inspecting the COBOL program file: " + cobolProgram);
         reset();
 
         // Use the statement locator to find the REPLACE statements in the COBOL program
@@ -112,6 +112,8 @@ public class Replace {
             replaceOn = true;
         }
         inspect_performed = true;
+        Log.info("Replace.inspectProgram(): Inspecting the COBOL program file: " + cobolProgram + " found "
+                + replaceMap.size() + " REPLACE statements");
     }
 
     /**
@@ -157,10 +159,16 @@ public class Replace {
     }
 
     public static String replaceInProgram(File program) {
-        // write the replaced program back to disk
+        // was there replace statements in the source, if not, return the original filename
+        if (!replaceOn) {
+            Log.debug("Replace.replaceInProgram(): No REPLACE statements found in the COBOL program file: " + program);
+            return program.getAbsolutePath();
+        }
 
-        String newFileName = program+"_MOD";
-        Log.warn("Replace.replaceInProgram(): Writing the COBOL program file: " + newFileName);
+        // write the replaced program back to disk
+        String newFileName = program.getAbsolutePath()+"_rpl";
+        boolean fileForReplacedSourceExisted = fileForReplacedSourceExists(newFileName);
+        Log.info("Replace.replaceInProgram(): Writing the COBOL program file: " + newFileName);
         try {
             BufferedWriter writer = new BufferedWriter(new FileWriter(newFileName));
             // read the program one line at the time
@@ -174,10 +182,35 @@ public class Replace {
             }
             writer.close();
             reader.close();
+
+            if (!fileForReplacedSourceExisted) {
+                updateFilePermissions(newFileName);
+            }
+
         } catch (IOException e) {
             Log.error("Replace.replaceInProgram(): Error writing the COBOL program file: " + program);
+            throw new RuntimeException("Replace.replaceInProgram(): Error writing the COBOL program file: " + program, e);
         }
         return newFileName;
+    }
+
+    /**
+     * Check if the file for the replaced source exists
+     * @param newFileName the name of the file to check
+     * @return true if the file exists, false otherwise
+     */
+    private static boolean fileForReplacedSourceExists(String newFileName) {
+        File file = new File(newFileName);
+        return file.exists();
+    }
+
+    /**
+     * Update the file permissions for the replaced source file
+     * @param newFileName the name of the file to update
+     */
+    private static void updateFilePermissions(String newFileName) {
+        String permissions = Config.getGeneratedFilesPermissionAll();
+        FilePermission.setFilePermissionForAllUsers(new File(newFileName), permissions);
     }
 
     public static void showReplaceSets() {
